Fix PS1 length issue
For some reason (unknown to me), the PS1 string from the previous commit
caused text going onto the next line to write over itself after reaching
a certain point, like can be observed if `\[` and `\]` are errantly
omitted from PS1 for color codes.

To work around this, the PS1 string is updated to be a two-line prompt
again.